backend/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
scikit-learn==1.2.2
sqlparse==0.4.4
gunicorn==20.1.0
jsonpickle==3.0.1
rdkit==2023.3.1
pillow==9.5.0
gTTS==2.3.2
